requirements.txt
numpy==1.24.3
tensorflow==2.11.0
scikit-learn==1.2.2
pandas==2.0.2
matplotlib==3.7.1
pyyaml==6.0
opencv-python==4.7.0.72
pillow==9.5.0
fastapi==0.98.0
uvicorn==0.22.0
requests==2.31.0